svsm/string: Changing string storage from char to u8
The string implementation currently uses the char type for backing but
actually stores utf-8 characters in each element in each way the string
constructed except when using push(). The only current usage of push()
is when reading byte I/O so from files or firmware.

Therefore, this patch changes the backing storage to u8 allowing it to
use utf-8 encoding and changes push() to also push utf-8 characters.
In addition, added the ability to append two string plus an initial set
of unit tests to check the new functionality.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>