task/schedule: Add a task runqueue for managing per CPU task switches
Implement an RB tree-based task list that orders the tree based
on the 'Runtime' of each task. The tree can then be used to determine
which task should run next when a task concedes execution on a
particular vCPU.

The runqueue task list includes a function named `schedule()` to
determine which task to run next as well as to update the tree and
internal task state in preparation for the switch. The actual code
to track and switch tasks for a particular vCPU are not part of this
commit.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>