task: Add termination handler for tasks
Previously if a task entry point exited the task handling code
would intentionally panic. This change adds support for short-
lived tasks where the task entry point returns.

When a task exits, the state of the task is set to TERMINATED
and the task is removed from the per-cpu scheduler tree. It
remains in the global task list (in a TERMINATED state) until
the task is closed.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>