svsm/module: Add basic support for loadable modules
Implement a module enumerator that locates and loads modules
from the RamFS and launches the module entry point. This is
the start of the work that will allow modules to be bundled in
the SVSM firmware image and executed at CPL-3 with a syscall
interface into the CPL-0 code.

At present, modules have very limited use. The intention of
this commit is to build the framework that will be expanded
on to support virtual memory allocation, CPL-3 entry/exit
and all the other support services required for seperation
of modules and the kernel.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>