svsm/mm: Add support for 16K bitmap allocator
The bitmap allocator is used to track virtual memory allocations.
Currently it is only used for perCPU virtual memory allocations
which are restricted to 1024 4k and 2M pages.

To support task-based virtual memory allocations we will need a larger
pool of pages. This patch adds a new 16K bitmap allocator that provides
that extra space and refactors the existing usage to abstract the
bitmap size into a type.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>