svsm/fs: Add map_view/unmap_view to access files using VA
The current filesystem is implemented as a RAM filesystem
therefore it makes sense to allow direct access to the file contents
without having to allocate and read files into a new buffer. To
achieve this, this commit adds two functions: map_view() and
unmap_view() which map the PAs containing the file contents
to a give virtual range within the pagetable for the given task.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>