svsm/task, svsm/module: Create tasks from memory-mapped ELF file
Currently, for a task to be created from an ELF file the file must be
entirely read into memory then passed to `create_elf_task()`. Now
that the file system supports mapping of files to virtual memory we
can update this function to take a file handle which prevents us from
having to copy a file contents into memory in order to create a task
from an ELF image.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>