cpu/gdt: Add GDT entries for supporting CPL-3
The existing GDT contains entries for kernel code at CPL-0.
In readiness for supporting loadable modules that running at
CPL-3 this adds entries in the GDT for 64 bit code and data selectors
with DPL=3.

The comments in the existing code reserved selectors for these
selectors, however in order to support the requirements for
syscall/sysret the selectors must be in a particular order as the
STAR MSR allows only setting of the 32-bit CPL-3 selector, implying the
64-bit values as offsets from this base. A comment has been added
to the code directing the reader to the relevant area of the AMD64
manual.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>