task/tasks: Extend Task structure with user-mode fields
Add a new structure UserTask which contains the configuration
and context of tasks that run with CPL-3. The presence of the
optional 'user' field can be used to distinguish between user
mode tasks running at CPL-3 and kernel tasks that only run at
CPL-0.

Only the structure is introduced at this point meaning that only
kernel tasks are supported with this commit. The implementation
of CPL-3 tasks will be added in later commits.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>