task: Initialise syscall/sysret and add basic handlers
For running tasks at CPL-3 we need some method to transition from
CPL-0 to CPL-3. The best way to achieve this transition is using
syscall/sysret.

This commit introduces a function which is used to setup the correct
configuration in syscall MSRs to allow the use of syscall/sysret for
switching between the different privilege levels.

In addition, a syscall handler has been implemented that adds support
for terminating a user-mode task, for allowing a context switch via the
kernel schedule() function and for a simple log function. These handlers
are only an initial implementation and likely to be reworked or replaced
in a subsequent commit.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>