fs/ramfs: Set pagetable permissions based on user/kernel tasks
When mapping a view of a RAM-FS file into virtual memory, the
required pagetable permission of the mapping is provided. To
support user-mode tasks running at CPL-3 we need to ensure that
the  permissions include the USER flag when configuring the
pagetable entries for these tasks.

This commit queries the provided task to see if it supports running
at CPL-3 and if it does it ensures the USER pagetable flag is set when
configuring pagetable permissions.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>