task: Allow passing of parameter to task entry point
Currently when creating a new task it is not possible to pass any
information directly to the task entry point. This makes it difficult
to provide context to a new task without resorting to some form
of global structure.

This commit adds a new parameter to create_kernel_task() that allows a
64-bit value to be passed directly to the task entry point.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>