task: create new tasks with interrupts disabled
The scheduler assumes that every new task initially executes with
interrupts disabled.  Thus, every new task needs to have its initial
stack frame created with interrupts disabled.  In addition, the initial
state of every new task should be consistent and not related to the
current flags disposition at the time the task is created, so the
initial task creation should use a constant flags value instead of the
current flags.

Signed-off-by: Jon Lange <jlange@microsoft.com>