kernel/idt: Save/restore GHCB MSR around HV event processing
HV event processing can use the GHCB in interrupt context. For example,
the restricted injection EOI handler on SEV-SNP issues a wrmsr on the
X2APIC which is handled via the GHCB. This alters the contents of the
GHCB MSR.

The SVSM switches to VMPL2 via switch_to_vmpl_unsafe using the GHCB MSR
protocol. This initiates the switch, then on return to VMPL0, checks the
contents of the MSR to determine whether the switch was successful or
not. If an HV is injected whilst in VMPL2, the contents of the MSR are
changed in the HV handler meaning that when the non-interrupt code path
resumes in VMPL0, the MSR contains an unexpected value and the SVSM
panics.

This patch saves the state of the GHCB MSR before processing HV events
and restores it afterwards to preserve the state during VMPL transfers.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>