kernel/cpu: Add module for configuring Local APIC
This adds a set of functions contained withing the struct HostApic that
is used to initialise and configure the Local APIC for the virtual
machine. This is separate to the Local APIC emulation implemented inside
the SVSM and can be used to configure the host to send interrupts to the
SVSM at VMPL0.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>