kernel: Initialise Local APIC and configure a periodic timer
This is for testing of interrupt handling and restricted injection only.

The X2APIC is configured to send a periodic timer interrupt at a very
slow period. This should be handled either by the IRQ20 handler or via
the HV handler depending on whether restricted injection is enabled or
not in the VMSA configured in the IGVM file.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>